fix(lib): move function from manage into lib
